+++ source/app/service-providers/fsal/fsal-directory.ts
@@ -54,7 +54,7 @@ const MARKDOWN_FILES = mdFileExtensions(true)
 const PROJECT_TEMPLATE: ProjectSettings = {
   // General values that not only pertain to the PDF generation
   title: 'Untitled', // Default project title is the directory's name
-  formats: [ 'html', 'chromium-pdf' ], // NOTE: Must correspond to the defaults in ProjectProperties.vue
+  profiles: [], // NOTE: Must correspond to the defaults in ProjectProperties.vue
   filters: [], // A list of filters (glob patterns) to exclude certain files
   cslStyle: '', // A path to an optional CSL style file.
   templates: {
@@ -345,7 +345,7 @@ export async function updateProjectProperties (dirObject: DirDescriptor, propert
 
   const titleUnchanged = dirObject._settings.project.title === properties.title
   const cslUnchanged = dirObject._settings.project.cslStyle === properties.cslStyle
-  const formatsUnchanged = JSON.stringify(dirObject._settings.project.formats) === JSON.stringify(properties.formats)
+  const formatsUnchanged = JSON.stringify(dirObject._settings.project.profiles) === JSON.stringify(properties.profiles)
   const filtersUnchanged = JSON.stringify(dirObject._settings.project.filters) === JSON.stringify(properties.filters)
   const templatesUnchanged = JSON.stringify(dirObject._settings.project.templates) === JSON.stringify(properties.templates)