+++ source/win-assets/SnippetsTab.vue
@@ -12,7 +12,7 @@
         v-bind:editable="true"
         v-on:select="currentItem = $event"
         v-on:add="addSnippet()"
-        v-on:remove="removeSnippet()"
+        v-on:remove="removeSnippet($event)"
       ></SelectableList>
     </template>
     <template #view2>
@@ -24,6 +24,7 @@
             v-model="currentSnippetText"
             v-bind:inline="true"
             v-bind:disabled="currentItem < 0"
+            v-on:confirm="renameSnippet()"
           ></TextControl>
           <ButtonControl
             v-bind:label="renameSnippetLabel"
@@ -207,11 +208,15 @@ export default defineComponent({
         .then(() => { this.updateAvailableSnippets(newName) })
         .catch(err => console.error(err))
     },
-    removeSnippet: function () {
+    removeSnippet: function (idx: number) {
+      if (idx > this.availableSnippets.length - 1 || idx < 0) {
+        return
+      }
+
       // Remove the current snippet.
       ipcRenderer.invoke('assets-provider', {
         command: 'remove-snippet',
-        payload: { name: this.availableSnippets[this.currentItem] }
+        payload: { name: this.availableSnippets[idx] }
       })
         .then(() => { this.updateAvailableSnippets() })
         .catch(err => console.error(err))
@@ -241,8 +246,8 @@ export default defineComponent({
      *
      * @return  {string}             The candidate's name, with a number suffix (-X) if necessary
      */
-    ensureUniqueName: function (candidate: string) {
-      if (this.availableSnippets.includes(candidate) === false) {
+    ensureUniqueName: function (candidate: string): string {
+      if (!this.availableSnippets.includes(candidate)) {
         return candidate // No duplicate detected
       }
 
@@ -252,7 +257,7 @@ export default defineComponent({
       if (match !== null) {
         // The candidate name already ends with a number-suffix --> extract it
         count = parseInt(match[1], 10)
-        candidate = candidate.substr(0, candidate.length - match[1].length - 1)
+        candidate = candidate.substring(0, candidate.length - match[1].length - 1)
       }
 
       while (this.availableSnippets.includes(candidate + '-' + String(count)) === true) {