+++ source/win-project-properties/App.vue
@@ -20,17 +20,17 @@
         v-bind:label="'Project Title'"
       ></TextControl>
 
-      <p v-if="selectedExportFormats.length === 0" class="warning">
+      <p v-if="selectedExportProfiles.length === 0" class="warning">
         <clr-icon shape="warning"></clr-icon>
         <!-- TODO: Translate! -->
-        <span>Please select at least one export format to build this project.</span>
+        <span>Please select at least one profile to build this project.</span>
       </p>
       <ListControl
         v-bind:label="exportFormatLabel"
         v-bind:model-value="exportFormatList"
-        v-bind:labels="[exportFormatUseLabel, exportFormatNameLabel]"
+        v-bind:labels="[exportFormatUseLabel, exportFormatNameLabel, 'Conversion']"
         v-bind:editable="[0]"
-        v-on:update:model-value="selectExportFormat($event)"
+        v-on:update:model-value="selectExportProfile($event)"
       ></ListControl>
     </div>
     <div
@@ -95,10 +95,12 @@ import TextControl from '@common/vue/form/elements/Text.vue'
 import { defineComponent } from 'vue'
 import { ProjectSettings } from '@dts/common/fsal'
 import { WindowTab } from '@dts/renderer/window'
+import { PandocProfileMetadata } from '@dts/common/assets'
+import { PANDOC_READERS, PANDOC_WRITERS, SUPPORTED_READERS } from '@common/util/pandoc-maps'
 
 const ipcRenderer = window.ipc
 
-interface ExportFormat { selected: boolean, format: string }
+interface ExportProfile { selected: boolean, name: string, conversion: string }
 
 export default defineComponent({
   components: {
@@ -110,8 +112,8 @@ export default defineComponent({
   data: function () {
     return {
       dirPath: '',
-      exportFormatMap: {} as { [key: string]: string },
-      selectedExportFormats: [ 'html', 'chromium-pdf' ], // NOTE: Must correspond to the defaults in fsal-directory.ts
+      profiles: [] as PandocProfileMetadata[],
+      selectedExportProfiles: [] as string[], // NOTE: Must correspond to the defaults in fsal-directory.ts
       patterns: [],
       cslStyle: '',
       texTemplate: '',
@@ -138,12 +140,17 @@ export default defineComponent({
     windowTitle: function (): string {
       return this.projectTitle
     },
-    exportFormatList: function (): ExportFormat[] {
-      // We need to return a list of { selected: boolean, format: 'string' }
-      return Object.keys(this.exportFormatMap).map(e => {
+    exportFormatList: function (): ExportProfile[] {
+      // We need to return a list of { selected: boolean, name: string, conversion: string }
+      return this.profiles.filter(e => SUPPORTED_READERS.includes(e.reader)).map(e => {
+        const reader = e.reader in PANDOC_READERS ? PANDOC_READERS[e.reader] : e.reader
+        const writer = e.writer in PANDOC_WRITERS ? PANDOC_WRITERS[e.writer] : e.writer
+        const conversionString = (e.isInvalid) ? 'Invalid' : [ reader, writer ].join(' → ')
+
         return {
-          selected: this.selectedExportFormats.includes(this.exportFormatMap[e]),
-          format: e
+          selected: this.selectedExportProfiles.includes(e.name),
+          name: this.getDisplayText(e.name),
+          conversion: conversionString
         }
       })
     },
@@ -164,7 +171,7 @@ export default defineComponent({
     }
   },
   watch: {
-    selectedExportFormats: function () {
+    selectedExportProfiles: function () {
       this.updateProperties()
     },
     projectTitle: function () {
@@ -188,21 +195,11 @@ export default defineComponent({
   },
   mounted: function () {
     // First, we need to get the available export formats
-    ipcRenderer.invoke('application', {
-      command: 'get-available-export-formats'
+    ipcRenderer.invoke('assets-provider', {
+      command: 'list-export-profiles'
     })
-      .then(exporterInformation => {
-        // We only need to know the readable string for an exportable format
-        // and the identifier. The list will be populated using the keys
-        // (human-readable string), and the actual value will consist of the
-        // values (the identifiers).
-        for (const info of exporterInformation) {
-          // NOTE: We are switching "id: readable" to "readable: id" here so
-          // that it's much easier to retrieve the identifier later on.
-          for (const key in info.formats) {
-            this.exportFormatMap[info.formats[key]] = key
-          }
-        }
+      .then((defaults: PandocProfileMetadata[]) => {
+        this.profiles = defaults
       })
       .catch(err => console.error(err))
 
@@ -221,19 +218,22 @@ export default defineComponent({
     })
   },
   methods: {
-    selectExportFormat: function (newListVal: ExportFormat[]) {
-      const newFormats = newListVal.filter(e => e.selected).map(e => {
-        return this.exportFormatMap[e.format]
-      })
-      this.selectedExportFormats = newFormats
+    selectExportProfile: function (newListVal: ExportProfile[]) {
+      const newProfiles = newListVal.filter(e => e.selected).map(e => {
+        return this.profiles.find(x => this.getDisplayText(x.name) === e.name)
+      }).filter(x => x !== undefined) as PandocProfileMetadata[]
+      this.selectedExportProfiles = newProfiles.map(x => x.name)
       this.updateProperties()
+    },
+    getDisplayText: function (name: string): string {
+      return name.substring(0, name.lastIndexOf('.'))
     },
     updateProperties: function () {
       ipcRenderer.invoke('application', {
         command: 'update-project-properties',
         payload: {
           properties: {
-            formats: this.selectedExportFormats.map(e => e), // De-proxy
+            profiles: this.selectedExportProfiles.map(e => e), // De-proxy
             filters: this.patterns.map(e => e), // De-proxy
             cslStyle: this.cslStyle,
             title: this.projectTitle,
@@ -254,7 +254,7 @@ export default defineComponent({
         .then(descriptor => {
           // Save the actually used formats.
           if (descriptor.project !== null) {
-            this.selectedExportFormats = descriptor.project.formats
+            this.selectedExportProfiles = descriptor.project.profiles
             this.patterns = descriptor.project.filters
             this.cslStyle = descriptor.project.cslStyle
             this.htmlTemplate = descriptor.project.templates.html